requirements.txt
asgiref==3.3.1
Django==3.1.7
python-slugify==4.0.1
pytz==2021.1
sqlparse==0.4.1
text-unidecode==1.3
